# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Trading Assistant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /notifications:
    get:
      operationId: getNotifications
      tags: [notifications]
      summary: Get notifications for current user
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Notifications list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationsList"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get admin statistics
      responses:
        "200":
          description: Admin stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/analyses:
    get:
      operationId: getAllAnalyses
      tags: [admin]
      summary: Get all analyses (admin only)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: All analyses
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysesList"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/notifications:
    post:
      operationId: broadcastNotification
      tags: [admin]
      summary: Broadcast notification to all users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BroadcastNotificationBody"
      responses:
        "201":
          description: Notification sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /superadmin/users:
    get:
      operationId: getAllUsers
      tags: [superadmin]
      summary: Get all users (superadmin only)
      responses:
        "200":
          description: All users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsersList"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createUser
      tags: [superadmin]
      summary: Create new user (superadmin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /superadmin/users/{id}:
    delete:
      operationId: deleteUser
      tags: [superadmin]
      summary: Delete user (superadmin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /superadmin/users/{id}/password:
    patch:
      operationId: resetUserPassword
      tags: [superadmin]
      summary: Reset user password (superadmin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetUserPasswordBody"
      responses:
        "200":
          description: Password reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /superadmin/users/{id}/role:
    patch:
      operationId: updateUserRole
      tags: [superadmin]
      summary: Update user role (superadmin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRoleBody"
      responses:
        "200":
          description: Role updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        displayName:
          type: string
        role:
          type: string
          enum: [user, admin, super_admin]
        selectedMode:
          type: string
          enum: [beginner, pro]
        themePreference:
          type: string
          enum: [light, dark]
        securityQuestion:
          type: string
        onboardingCompleted:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - displayName
        - role
        - selectedMode
        - themePreference
        - onboardingCompleted
        - createdAt
...
    Analysis:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        instrument:
          type: string
        timeframe:
          type: string
        userInputContext:
          type: string
          nullable: true
        mode:
          type: string
          enum: [beginner, pro]
        validUntil:
          type: string
          format: date-time
        marketCondition:
          type: string
          nullable: true
        riskLevel:
          type: string
          nullable: true
        confidenceMin:
          type: integer
          nullable: true
        confidenceMax:
          type: integer
          nullable: true
        mainScenario:
          type: string
          nullable: true
        alternativeScenario:
          type: string
          nullable: true
        whyReason:
          type: string
          nullable: true
        failureConditions:
          type: string
          nullable: true
        baseCase:
          type: string
          nullable: true
        bullishScenario:
          type: string
          nullable: true
        bearishScenario:
          type: string
          nullable: true
        keyDriversTechnical:
          type: string
          nullable: true
        keyDriversFundamental:
          type: string
          nullable: true
        marketContext:
          type: string
          nullable: true
        invalidationConditions:
          type: string
          nullable: true
        uncertaintyNotes:
          type: string
          nullable: true
        feedback:
          $ref: "#/components/schemas/Feedback"
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - instrument
        - timeframe
        - mode
        - validUntil
        - createdAt
    AnalysesList:
      type: object
      properties:
        analyses:
          type: array
          items:
            $ref: "#/components/schemas/Analysis"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - analyses
        - total
        - page
        - limit
...
    Feedback:
      type: object
      properties:
        id:
          type: integer
        analysisId:
          type: integer
        feedbackType:
          type: string
          enum: [useful, not_useful]
        outcome:
          type: string
          enum: [correct, wrong, unknown]
          nullable: true
        note:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - analysisId
        - feedbackType
        - createdAt
...
    Notification:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
          nullable: true
        targetRole:
          type: string
          nullable: true
        title:
          type: string
        message:
          type: string
        type:
          type: string
          enum: [info, warning, error]
        readAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - message
        - type
        - createdAt
    NotificationsList:
      type: object
      properties:
        notifications:
          type: array
          items:
            $ref: "#/components/schemas/Notification"
        unreadCount:
          type: integer
      required:
        - notifications
        - unreadCount
    BroadcastNotificationBody:
      type: object
      properties:
        title:
          type: string
        message:
          type: string
        type:
          type: string
          enum: [info, warning, error]
          default: info
        targetRole:
          type: string
          enum: [user, admin, super_admin]
          nullable: true
      required:
        - title
        - message
    AdminStats:
      type: object
      properties:
        totalUsersToday:
          type: integer
        totalAnalysesToday:
          type: integer
        totalAnalysesThisWeek:
          type: integer
        totalAnalysesThisMonth:
          type: integer
        totalUsers:
          type: integer
        instrumentBreakdown:
          type: array
          items:
            type: object
            properties:
              instrument:
                type: string
              count:
                type: integer
            required:
              - instrument
              - count
        modeBreakdown:
          type: object
          properties:
            beginner:
              type: integer
            pro:
              type: integer
          required:
            - beginner
            - pro
      required:
        - totalUsersToday
        - totalAnalysesToday
        - totalAnalysesThisWeek
        - totalAnalysesThisMonth
        - totalUsers
        - instrumentBreakdown
        - modeBreakdown
    UserWithStats:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        displayName:
          type: string
        role:
          type: string
          enum: [user, admin, super_admin]
        selectedMode:
          type: string
          enum: [beginner, pro]
        analysisCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - displayName
        - role
        - selectedMode
        - analysisCount
        - createdAt
    UsersList:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/UserWithStats"
        total:
          type: integer
      required:
        - users
        - total
    CreateUserBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        displayName:
          type: string
        role:
          type: string
          enum: [user, admin, super_admin]
          default: user
        securityQuestion:
          type: string
          default: "Nama hewan peliharaan pertama kamu?"
        securityAnswer:
          type: string
          default: "default"
      required:
        - email
        - password
        - displayName
    ResetUserPasswordBody:
      type: object
      properties:
        newPassword:
          type: string
          minLength: 6
      required:
        - newPassword
    UpdateUserRoleBody:
      type: object
      properties:
        role:
          type: string
          enum: [user, admin, super_admin]
      required:
        - role